Partial connection to Kscope. Start implementing mechanizm for automated path conversion.

+++ src/ssh/connect/SSHclient.java
@@ -1,14 +1,17 @@
 package ssh.connect;
 
-
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;  
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.net.InetAddress;
-import java.net.NetworkInterface;
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Properties;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -28,9 +31,11 @@ import com.trilead.ssh2.Session;
 import com.trilead.ssh2.StreamGobbler;
 
 
-
 /**
- * Ver.0.13 
+ * Ver.0.16
+ * 
+ *  Connecting to K-scope
+ *  
  * Orion SSH + JSch
  * Parsing Makefiles for replacement patterns
  * Unique temporary directory names
@@ -41,31 +46,75 @@ import com.trilead.ssh2.StreamGobbler;
 
 public class SSHclient {
 
+	static String searchable_files_extensions = "f90,";  // extensions of files to search for absolute paths
+	
 	/**
 	 * @param args
 	 * @throws JSchException 
 	 * @throws SftpException 
+	 * @throws NoSuchAlgorithmException 
 	 */
-	public static void main(String[] args) throws IOException, JSchException, SftpException {
-		SSHbasic basic_coonection = new SSHbasic();
-		basic_coonection.makeConnect();
+	public static void main(String[] args) throws IOException, JSchException, SftpException, NoSuchAlgorithmException {
+		SSHbasic basic_connection = new SSHbasic();
+		basic_connection.local_path = args[args.length-1];
+		for (String arg: args) {
+			System.out.println(arg);
+		}
+		detectPaths(basic_connection.local_path);
+		//basic_connection.makeConnect();
+	}
+	
+	/**
+	 * Detect paths in makfiles and source files
+	 * @param local_path local path to start search
+	 */
+	private static List<File> detectPaths(String local_path) {
+		List<File> files_to_check = new ArrayList<File>();
+		files_to_check = getFilesList(files_to_check, local_path,searchable_files_extensions);
+		for (File file:files_to_check) {
+			String s = file.toString();
+			/// look for absolute paths
+		}
+		return files_to_check;
+	}
+
+	/**
+     * Get list of files with extensions in comma-separated list.
+     * Search in path including subdirectories.
+     * 
+     * @param files	List of files found before
+     * @param start_path Path to start search
+     * @param extensions Comma-separated list of extensions
+     * @return
+     */
+	private static List<File> getFilesList(List<File> files, String start_path, String extensions) {
+		
+	
+
+		return null;
 	}
 
 	static class SSHbasic {
-		//String endLineStr = " # "; // it is dependent to the server  
+		
+		// Remote path to F_Front and atool  
+		String Ffront_path = ""; 
 		
 		// Initializing parameters with default values
-	    String host = "192.168.56.101"; // host IP  
-	    String user = "peter"; // username for SSH connection  
-	    String password = "simsim"; // password for SSH connection  
+	    String host = ""; // host IP  
+	    String user = ""; // username for SSH connection  
+	    String password = ""; // password for SSH connection  
+	    int group_id; // user group ("kscope") ID	   
+	    int port; // default SSH port  
 	    
 		// local project folder. Must contain Makefile and all files necessary for building.  
-	    String local_path = "/Users/peterbryzgalov/work/HimenoBMT_Dprof"; 
-	    String remote_path = "/home/peter/kscope/"; 
-	    String remote_full_path = ""; // path including temporary directory and archive name without extention
+	    String local_path = ""; 
+	    String remote_path = ""; 
+	    String remote_full_path = ""; // path including temporary directory and archive name without extension
 	    String archive = "";
 	    
 	    String makefiles = "";
+	    String make = "";
+	    String make_options = "";
 	    static private final Pattern placeholder_pattern = Pattern.compile("<([\\w\\d\\-_]*)>");
 	    private String default_archive_filename = "archive.zip"; 
 	    
@@ -79,8 +128,6 @@ public class SSHclient {
 	    // Exclude selected file types.
 	    String file_filter = ".*,*.tar,*.html,*.zip,*.jpg.*.orgin";
 	    	    
-	    int group_id = 1002; // kscope group ID	   
-	    int port = 22; // default SSH port  
 	    
 	    // JSch parameters initialization
 	    com.jcraft.jsch.Session session = null;
@@ -92,168 +139,214 @@ public class SSHclient {
     	String[] makefiles_list = null;
 	    
 	    
-	    public SSHbasic()  throws IOException  {	
+	    public SSHbasic()  throws IOException, NoSuchAlgorithmException  {	
 	    	
 	    	// Read parameters from configuration file
     		Properties prop = new Properties();
         	try {
                 //load properties from config file
         		prop.load(new FileInputStream("config.txt"));
-        		user = prop.getProperty("user").replaceAll("\\s","");
-        		password = prop.getProperty("password").replaceAll("\\s","");
-        		host = prop.getProperty("host").replaceAll("\\s","");
-        		port = Integer.parseInt(prop.getProperty("port").replaceAll("\\s",""));
-        		local_path = prop.getProperty("local_path").replaceAll("\\s","");
-        		remote_path = prop.getProperty("remote_path").replaceAll("\\s","");     
-        		file_filter = prop.getProperty("file_filter").replaceAll("\\s","")+",*.origin";
-        		makefiles = prop.getProperty("makefiles").replaceAll("\\s","");
-        		if (makefiles.length() > 0) makefile_filenames = makefiles.split(",");
+        		user = updateProperty(prop, "user");
+        		password = updateProperty(prop, "password");
+        		try {
+        			group_id = Integer.parseInt(updateProperty(prop, "group_id"));
+        		} catch (NumberFormatException e) {
+        			group_id = 1022;
+        			System.out.println("Default groupID 1022 used.");
+        		}
+        		host = updateProperty(prop, "host");
+        		try {
+        			port = Integer.parseInt(updateProperty(prop, "port"));
+        		} catch (NumberFormatException e) {
+        			port = 22;
+        			System.out.println("Default port 22 used.");
+        		}
+        		// local_path = updateProperty(prop, "local_path");
+        		remote_path = updateProperty(prop, "remote_path");     
+        		
+        		// *.origin - reserved for original copies of edited make files.
+        		file_filter = updateProperty(prop, "file_filter")+",*.origin";
+        		makefiles = updateProperty(prop, "makefiles");
+        		make = updateProperty(prop,"make");
+        		make_options = prop.getProperty("make_options"); // no need to remove spaces
+        		if (make_options == null) make_options = "";
+        		Ffront_path = updateProperty(prop,"Ffront_path");
+        		if (makefiles != null && makefiles.length() > 0) makefile_filenames = makefiles.split(",");
         	} catch (IOException e) {
         		e.printStackTrace();
         		System.out.println("Default parameters used.");
         	}
 
         	// Remote tmp directory name generation
-        	String mac = getMacAddress();
-        	tmp_dir = String.format("tmp-%s-%d-%s", System.getProperty("user.name"),System.currentTimeMillis()/1000,mac);
-
+        	tmp_dir = String.format("tmp%d_%s", System.currentTimeMillis()/1000, getTmpDirName(System.getProperty("user.name")));
         	remote_tmp = remote_path + "/" + tmp_dir;
         	remote_tmp = remote_tmp.replaceAll("//", "/");	
         	//System.out.println(remote_tmp);
     	}
     	
-	    private String getMacAddress() {
-	    	InetAddress ip;
-	    	StringBuilder sb = null;
-	    	try {
-	     
-	    		ip = InetAddress.getLocalHost();
-	    		//System.out.println("Current IP address : " + ip.getHostAddress());
-	     
-	    		NetworkInterface network = NetworkInterface.getByInetAddress(ip);
-	     
-	    		byte[] mac = network.getHardwareAddress();
-	     
-	    		//System.out.print("Current MAC address : ");
-	     
-	    		sb = new StringBuilder();
-	    		for (int i = 0; i < mac.length; i++) {
-	    			sb.append(String.format("%02X", mac[i]));		
-	    		}
-
-	    	} catch (Exception e){
-
-	    		e.printStackTrace();
-
-	    	}
-	    	String macS = sb.toString();
-	    	//macS = macS.replaceAll("-", "").replaceAll("00","");
-	    	return macS;
+	    /**
+	     * Return hashcode of a string, adopted to use as a filename part
+	     * @param name - some string
+	     * @return hashcode
+	     * @throws UnsupportedEncodingException 
+	     * @throws NoSuchAlgorithmException 
+	     */
+		private Object getTmpDirName(String name) throws UnsupportedEncodingException, NoSuchAlgorithmException {
+			byte[] tmp_dir_name_bytes = name.getBytes("UTF-8");
+        	MessageDigest md = MessageDigest.getInstance("SHA-1");
+        	tmp_dir_name_bytes = md.digest(tmp_dir_name_bytes);
+        	String hash = URLEncoder.encode(tmp_dir_name_bytes.toString(),"UTF-8");
+        	hash = hash.replaceAll("[^\\w\\d-_]","");
+        	return hash;
 		}
 
+
+		/**
+		 * Return property value without space characters.
+		 * @param prop Property set.
+		 * @param property_name Name of a property.
+		 * @return property value without space characters
+		 */
+		private String updateProperty(Properties prop, String property_name) {
+			String property = prop.getProperty(property_name);
+			if (property == null) return "";
+			property = property.replaceAll("\\s","");
+			return property;
+		}
+
+
+		/**
+		 * Method for compiling source code on remote machine.
+		 * 
+		 * 1.Connect to remote machine
+		 * 2. Create archive with source files,
+		 * 3. Upload archive to temporary directory
+		 * 4. Extract source files from archive on remote machine
+		 * 5. Execute Make command
+		 * 6. Pick up xml files
+		 * 7. Download xml files to local source files location
+		 * 8. Remove remote temporary directory 
+		 * 
+		 * @throws IOException
+		 * @throws JSchException
+		 * @throws SftpException
+		 */
 		public void makeConnect()  throws IOException, JSchException, SftpException {
-	    	// JSch connect and upload
-	    	
-	    	// get a new session    
-	        session = new_session();  
-	    	channel=session.openChannel("sftp");
-	    	channel.connect();
-	        sftp_channel=(ChannelSftp)channel;
-	       
-	    	createArchiveAndUpload();
-	    	
-	    	// Execute remote commands
-	    	Connection conn = new Connection(host,port);
-	    	conn.connect();
+			
+			// Orion SSH
+			Connection conn = new Connection(host,port);
 
-	    	boolean isAuthenticated = conn.authenticateWithPassword(user, password);
+			try {
 
-	    	if (isAuthenticated == false)
-	    		throw new IOException("Authentication failed.");
-	    	Session sess = conn.openSession();
-	    	sess.execCommand("export PATH=$PATH:/home/peter/kscope && echo $PATH && cd "+remote_tmp+" && pwd && unzip -o "+archive+" && cd "+noExtension(archive)+" && make");
-	    	System.out.println("Here is some information about the remote host:");
-	    	InputStream stdout = new StreamGobbler(sess.getStdout());
-	    	InputStream stderr = new StreamGobbler(sess.getStderr());
-	    	BufferedReader br = new BufferedReader(new InputStreamReader(stdout));
-	    	BufferedReader stderrReader = new BufferedReader(new InputStreamReader(stderr));
-	    	
-	    	try {
-	    		while (true)
-	    		{
-	    			String line = br.readLine();
-	    			if (line == null)
-	    				break;
-	    			System.out.println(line);
-	    		}
-	    		while (true)
-				{
-					String line = stderrReader.readLine();
-					if (line == null)
-						break;
-					System.err.println(line);
-				}	    		
-	    	} finally {
-	    		System.out.println("ExitCode: " + sess.getExitStatus());
-	    		System.out.println("Connction:" +conn.getConnectionInfo());
-	    		br.close();
-	    		stderrReader.close();
-	    		sess.close();
-	    		
-	    	}
-	    	
-	    	sess = conn.openSession();
-	    	sess.execCommand("cd "+remote_full_path+" && find -name \"*.xml\"");
-	    	stdout = new StreamGobbler(sess.getStdout());
-	    	br = new BufferedReader(new InputStreamReader(stdout));
-	    	StringBuilder response = new StringBuilder();
-	    	try {
-	    		while (true)
-	    		{
-	    			String line = br.readLine();
-	    			if (line == null)
-	    				break;
-	    			response.append(line+"\n\r");
-	    			//System.out.println(line);
-	    		}	    		    		
-	    	} finally {
-	    		br.close();
-	    		sess.close();	    		
-	    	}	    	
-	    	
-	    	String[] filenames = response.toString().replaceAll("(\\s\\./)|(^\\./)", "").split("\n");
-	    	System.out.println("\n\n--------");
-	    	
-	    	// Download XML files with JSch	        
-	    	for (String filename:filenames) {
-	    		if (filename.length() < 2) continue;
-	    		String remote_filename =remote_full_path+"/"+filename;
-	    		String local_filename = local_path+"/" + filename;
-	        	//System.out.println("Downloading "+remote_filename+" to " + local_filename);
-	        	try {
-	        	File local_file = new File(local_filename);
-	        	if (!local_file.exists()) {
-	        		local_file.getParentFile().mkdirs();
-	        		//local_file.createNewFile();
-	        	}
-	        	sftp_channel.get(remote_filename, local_filename, monitor, mode);	  
-	        	} catch (Exception e) {
-	        		e.printStackTrace(System.err);
-	        	}
-	    	}
-	    	
-	    	//TODO remove remote temporary directory and archive
-	    	
-	    	//orionSSH
-	    	conn.close();
-	    	
-	    	//JSch
-	    	sftp_channel.exit();
-		    channel.disconnect();  
-	        session.disconnect();
-	        
-	    	System.out.println("finished.");
-	        System.exit(0);
+				// 1.
+				// JSch connect and upload
+				// get a new session    
+				session = new_session();  
+				channel=session.openChannel("sftp");
+				channel.connect();
+				sftp_channel=(ChannelSftp)channel;
+
+				// 2. 3.  
+				createArchiveAndUpload();
+
+				// Execute remote commands
+
+				conn.connect();
+
+				boolean isAuthenticated = conn.authenticateWithPassword(user, password);
+
+				if (isAuthenticated == false)
+					throw new IOException("Authentication failed.");
+				Session sess = conn.openSession();
+				// 4. 5.
+				sess.execCommand("export PATH=$PATH:"+Ffront_path+" && echo $PATH && cd "+remote_tmp+" && pwd && unzip -o "+archive+" && cd "+noExtension(archive)+" && " + make + " " + make_options);
+				InputStream stdout = new StreamGobbler(sess.getStdout());
+				InputStream stderr = new StreamGobbler(sess.getStderr());
+				BufferedReader br = new BufferedReader(new InputStreamReader(stdout));
+				BufferedReader stderrReader = new BufferedReader(new InputStreamReader(stderr));
+				// Read and display output
+				try {
+					while (true)
+					{
+						String line = br.readLine();
+						if (line == null)
+							break;
+						System.out.println(line);
+					}
+					while (true)
+					{
+						String line = stderrReader.readLine();
+						if (line == null)
+							break;
+						System.err.println(line);
+					}	    		
+				} finally {
+					System.out.println("ExitCode: " + sess.getExitStatus());
+					System.out.println("Connction:" +conn.getConnectionInfo());
+					br.close();
+					stderrReader.close();
+					sess.close();
+
+				}
+
+				sess = conn.openSession();
+				// 6. Pick up xml files
+				sess.execCommand("cd "+remote_full_path+" && find -name \"*.xml\"");
+				stdout = new StreamGobbler(sess.getStdout());
+				br = new BufferedReader(new InputStreamReader(stdout));
+				StringBuilder response = new StringBuilder();
+				try {
+					while (true)
+					{
+						String line = br.readLine();
+						if (line == null)
+							break;
+						response.append(line+"\n\r");
+						//System.out.println(line);
+					}	    		    		
+				} finally {
+					br.close();
+					sess.close();	    		
+				}	    	
+
+				String[] filenames = response.toString().replaceAll("(\\s\\./)|(^\\./)", "").split("\n");
+				System.out.println("\n--------");
+
+				// 7. Download XML files with JSch	        
+				for (String filename:filenames) {
+					if (filename.length() < 2) continue;
+					String remote_filename =remote_full_path+"/"+filename;
+					String local_filename = local_path+"/" + filename;
+					//System.out.println("Downloading "+remote_filename+" to " + local_filename);
+					try {
+						File local_file = new File(local_filename);
+						if (!local_file.exists()) {
+							local_file.getParentFile().mkdirs();
+							//local_file.createNewFile();
+						}
+						sftp_channel.get(remote_filename, local_filename, monitor, mode);	  
+					} catch (Exception e) {
+						e.printStackTrace(System.err);
+					}
+				}
+
+				// 8.
+				//TODO remove remote temporary directory and archive
+
+			} catch (Exception ex) {
+				ex.printStackTrace();
+			} finally {
+				//orionSSH
+				conn.close();
+
+				//JSch
+				sftp_channel.exit();
+				channel.disconnect();  
+				session.disconnect();
+
+				System.out.println("finished.");
+				System.exit(0);
+			}
 	    } 	
 	    
 	    /**
@@ -296,16 +389,18 @@ public class SSHclient {
 		    System.out.println("Local directory: " + sftp_channel.lpwd());
 		     
 		    // Parse make files
-		    makefiles_list = getFilesList("", local_path, makefiles).split(",");
-		    for (String makefile_path : makefiles_list) {
-		    	// Save original files
-		    	File makefile_org = new File(makefile_path);
-		    	File makefile_backup = new File(makefile_path+".origin");
-		    	FileUtils.copyFile(makefile_org, makefile_backup);
-		    	String s = FileUtils.readFileToString(makefile_org,"UTF-8");
-		    	
-		    	s = replacePlaceholdersInMakefile(s);
-		    	saveString2File(s,makefile_path);
+		    if (makefiles.length() > 0) {
+		    	makefiles_list = getFilesList("", local_path, makefiles).split(",");
+		    	for (String makefile_path : makefiles_list) {
+		    		// Save original files
+		    		File makefile_org = new File(makefile_path);
+		    		File makefile_backup = new File(makefile_path+".origin");
+		    		FileUtils.copyFile(makefile_org, makefile_backup);
+		    		String s = FileUtils.readFileToString(makefile_org,"UTF-8");
+
+		    		s = replacePlaceholdersInMakefile(s);
+		    		saveString2File(s,makefile_path);
+		    	}
 		    }
 		    
 		    // Create ZIP archive
@@ -324,11 +419,13 @@ public class SSHclient {
 		    file.delete();
 		    
 		    // Restore original makefiles
-		    for (String makefile_path : makefiles_list) {
-		    	File makefile_org = new File(makefile_path);
-		    	File makefile_backup = new File(makefile_path+".origin");
-		    	FileUtils.copyFile(makefile_backup, makefile_org);
-		    	makefile_backup.delete();
+		    if (makefiles.length() > 0) {
+		    	for (String makefile_path : makefiles_list) {
+		    		File makefile_org = new File(makefile_path);
+		    		File makefile_backup = new File(makefile_path+".origin");
+		    		FileUtils.copyFile(makefile_backup, makefile_org);
+		    		makefile_backup.delete();
+		    	}
 		    }
 		}
